requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
future==0.17.1
gast==0.2.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
mock==2.0.0
mpmath==1.1.0
numpy==1.16.2
pbr==5.1.3
promise==2.2.1
protobuf==3.7.0
pyparsing==2.3.1
python-dateutil==2.8.0
requests==2.21.0
scipy==1.2.1
six==1.12.0
sympy==1.4
tensorboard==1.13.1
tensorflow-datasets==1.0.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-metadata==0.13.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.2
Werkzeug==0.14.1
wrapt==1.11.1
